Workaround for ballooning configs; Fix overviewpage.cpp newlines
Ballooning configs are because whenever I wrote this stuff I was stupid
and comments for the whole preset are stripped, but they're added back
onto every style, which means there's exponential growth in size.

For now I'm only adding comments to the first style, but of course that's
very much less than ideal. The only reason I do this is because a proper
fix is much more difficult, and I'll probably break other things. I don't
like this code and I don't think it's worthwhile to fix since V2 already
handles this properly (and much better).

Overview got messed up with the newlines, so I fixed that, and removed
the excessive usage of `std::endl`...