PCUI::Notifier purge pending events on rebind
If a proxy is unbound, then bound to the previous data, it was possible
for an in-flight event from pre-unbind to hit the notifier.

If that occurred, the data would match, so the event would be processed
by the notifier, and mInFlight would be decremented past zero.

It's possibly worth making sure mInFlight is a sane value on the
notifier side too, but purging events on rebind stops (at least this
particular) problem at the source.