Config: Import bladestyle formatted; close config after import
The latter is to ensure the bladestyles are formatted on first open
after import, but semantically in this system the import shouldn't keep
the config open anyways.

The style formatting code we'll call a preview. It should really be its
own separate function/feature, but I'm not messing with it right now. I
can't handle any more scope creep