Utils: remove getDPIScaleFactor; Maybe fix bmp scaling?
Image::getDPIScaleFactor was always a stub anyways. The sized version of
loadPNG is more reliable for getting a particular sized image, and
loadPNG is primarily just for wxStaticBitmap which sets its own size
anyways, so it probably doesn't need to set scale factor anyways, but it
will for now.

Until bitmaps are annoying again (virtually guaranteed)