Update utils/string
Add a generic trim to be used where the Cpp trim was erroneously being
overused (number mode for it was particularly not expected for many of
its uses that weren't, well, cpp names).

Removed templates and moved everything into `.cpp`. wxString isn't used
directly with it anymore.

Misc related updates